pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "conidk"
version = "1.0.0"
authors = [
  { name="Thiago Alberto", email="talberto@google.com" },
]
description = "A high-level API for Google's Conversational Insights."
readme = "README.md"
license = { file="LICENSE.txt" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "google-auth",
    "google-api-python-client",
    "google-cloud-contact-center-insights",
    "google-cloud-speech",
    "google-cloud-dlp",
    "google-cloud-storage",
    "vertexai",
    "jsonschema",
    "strenum",
    "google-cloud-texttospeech",
    "pydub",
    "requests",
    "google-cloud-dialogflow",
    "google-cloud-dialogflow-cx",
    "pandas",
    "gspread",
    "google-cloud-iam",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint",
    "coverage",
    "pytest-cov",
    "mypy",
    "pandas-stubs",
    "pytest-mock",
]
test = [
    "pytest",
]

[tool.mypy]
mypy_path = "src"
ignore_missing_imports = true

# Ignore missing stubs for specific third-party libraries
[[tool.mypy.overrides]]
module = [
  "conidk.*",
  "google.*",
  "gspread",
  "pydub",
  "pytest",
  "requests",
  "unittest.*",
  "vertexai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = [
  "src",
]

[tool.setuptools.package-data]
"conidk.workflow.utils.schemas" = ["*.json"]
"conidk.workflow.utils.defaults" = ["*.json"]

[tool.pylint.main]
source-roots = ["src"]